Update to InstallShield config
Update to InstallShield config to build installer using .Net 2.0 instead
of .Net 4.6.  This file is actually generated at build time (and hence
is over-written anyway) but hopefully uploading it to GitHub will act as
a reminder to set the InstallShield options to use .Net 2.0 because
using .Net 4.0 will cause 4.0 to become 4.6 - this then causes an
install error 1001.